[OpenSSL] Fully merge openssl3 patch.
Fully merge the openssl3 patch originally provided by @ellert in c07db3cf.
