+++ cmake/XRootDDefaults.cmake
@@ -26,5 +26,4 @@ option( PYPI_BUILD       "The project is being built for PyPI release"          
 option( ENABLE_VOMS      "Enable VOMS plug-in if possible."                               TRUE )
 option( ENABLE_XRDEC     "Enable erasure coding component."                               FALSE )
 option( ENABLE_ASAN      "Enable adress sanitizer."                                       FALSE )
-option( ENABLE_OPENSSL3  "Enable openssl3 support."                                       TRUE )
 define_default( XRD_PYTHON_REQ_VERSION 2.4 )
