+++ src/XrdCrypto.cmake
@@ -15,35 +15,19 @@ set( XRD_CRYPTO_SOVERSION 2 )
 set( XRD_CRYPTO_LITE_VERSION   2.0.0 )
 set( XRD_CRYPTO_LITE_SOVERSION 2 )
 
-if( ENABLE_OPENSSL3 )
-  set( XrdCryptoSources
-    XrdCrypto/XrdCryptoAux.cc                 XrdCrypto/XrdCryptoAux.hh
-    XrdCrypto/XrdCryptoTrace.hh
-    XrdCrypto/XrdCryptoBasic.cc               XrdCrypto/XrdCryptoBasic.hh
-    XrdCrypto/XrdCryptoCipher.cc              XrdCrypto/XrdCryptoCipher.hh
-    XrdCrypto/XrdCryptoFactory.cc             XrdCrypto/XrdCryptoFactory.hh
-    XrdCrypto/XrdCryptoMsgDigest.cc           XrdCrypto/XrdCryptoMsgDigest.hh
-    XrdCrypto/XrdCryptoRSA.cc                 XrdCrypto/XrdCryptoRSA.hh
-    XrdCrypto/XrdCryptoX509.cc                XrdCrypto/XrdCryptoX509.hh
-    XrdCrypto/openssl3/XrdCryptoX509Chain.cc  XrdCrypto/XrdCryptoX509Chain.hh
-    XrdCrypto/XrdCryptoX509Crl.cc             XrdCrypto/XrdCryptoX509Crl.hh
-    XrdCrypto/XrdCryptoX509Req.cc             XrdCrypto/XrdCryptoX509Req.hh
-    XrdCrypto/XrdCryptogsiX509Chain.cc        XrdCrypto/XrdCryptogsiX509Chain.hh )
-else()
-  set( XrdCryptoSources
-    XrdCrypto/XrdCryptoAux.cc               XrdCrypto/XrdCryptoAux.hh
-    XrdCrypto/XrdCryptoTrace.hh
-    XrdCrypto/XrdCryptoBasic.cc             XrdCrypto/XrdCryptoBasic.hh
-    XrdCrypto/XrdCryptoCipher.cc            XrdCrypto/XrdCryptoCipher.hh
-    XrdCrypto/XrdCryptoFactory.cc           XrdCrypto/XrdCryptoFactory.hh
-    XrdCrypto/XrdCryptoMsgDigest.cc         XrdCrypto/XrdCryptoMsgDigest.hh
-    XrdCrypto/XrdCryptoRSA.cc               XrdCrypto/XrdCryptoRSA.hh
-    XrdCrypto/XrdCryptoX509.cc              XrdCrypto/XrdCryptoX509.hh
-    XrdCrypto/XrdCryptoX509Chain.cc         XrdCrypto/XrdCryptoX509Chain.hh
-    XrdCrypto/XrdCryptoX509Crl.cc           XrdCrypto/XrdCryptoX509Crl.hh
-    XrdCrypto/XrdCryptoX509Req.cc           XrdCrypto/XrdCryptoX509Req.hh
-    XrdCrypto/XrdCryptogsiX509Chain.cc      XrdCrypto/XrdCryptogsiX509Chain.hh )
-endif()
+set( XrdCryptoSources
+     XrdCrypto/XrdCryptoAux.cc               XrdCrypto/XrdCryptoAux.hh
+     XrdCrypto/XrdCryptoTrace.hh
+     XrdCrypto/XrdCryptoBasic.cc             XrdCrypto/XrdCryptoBasic.hh
+     XrdCrypto/XrdCryptoCipher.cc            XrdCrypto/XrdCryptoCipher.hh
+     XrdCrypto/XrdCryptoFactory.cc           XrdCrypto/XrdCryptoFactory.hh
+     XrdCrypto/XrdCryptoMsgDigest.cc         XrdCrypto/XrdCryptoMsgDigest.hh
+     XrdCrypto/XrdCryptoRSA.cc               XrdCrypto/XrdCryptoRSA.hh
+     XrdCrypto/XrdCryptoX509.cc              XrdCrypto/XrdCryptoX509.hh
+     XrdCrypto/XrdCryptoX509Chain.cc         XrdCrypto/XrdCryptoX509Chain.hh
+     XrdCrypto/XrdCryptoX509Crl.cc           XrdCrypto/XrdCryptoX509Crl.hh
+     XrdCrypto/XrdCryptoX509Req.cc           XrdCrypto/XrdCryptoX509Req.hh
+     XrdCrypto/XrdCryptogsiX509Chain.cc      XrdCrypto/XrdCryptogsiX509Chain.hh )
 
 #-------------------------------------------------------------------------------
 # The XrdCrypto library
@@ -70,15 +54,9 @@ set_target_properties(
 # The XrdCryptoLite library
 #-------------------------------------------------------------------------------
 
-if( ENABLE_OPENSSL3 )
-  set( XrdCryptoLiteSources
-       XrdCrypto/XrdCryptoLite.cc   XrdCrypto/XrdCryptoLite.hh
-       XrdCrypto/openssl3/XrdCryptoLite_bf32.cc )
-else()
-  set( XrdCryptoLiteSources
-       XrdCrypto/XrdCryptoLite.cc   XrdCrypto/XrdCryptoLite.hh
-       XrdCrypto/XrdCryptoLite_bf32.cc )
-endif()
+set( XrdCryptoLiteSources
+     XrdCrypto/XrdCryptoLite.cc      XrdCrypto/XrdCryptoLite.hh
+     XrdCrypto/XrdCryptoLite_bf32.cc )
 
 add_library(
   XrdCryptoLite
@@ -110,31 +88,17 @@ set_target_properties(
 if( BUILD_CRYPTO )
   include_directories( ${OPENSSL_INCLUDE_DIR} )
 
-  if( ENABLE_OPENSSL3 )
-    set( XrdCryptosslSources 
-         XrdCrypto/openssl3/XrdCryptosslAux.cc       XrdCrypto/XrdCryptosslAux.hh
-         XrdCrypto/openssl3/XrdCryptosslgsiAux.cc
-         XrdCrypto/openssl3/XrdCryptosslCipher.cc    XrdCrypto/openssl3/XrdCryptosslCipher.hh
-         XrdCrypto/XrdCryptosslMsgDigest.cc          XrdCrypto/XrdCryptosslMsgDigest.hh
-         XrdCrypto/openssl3/XrdCryptosslRSA.cc       XrdCrypto/XrdCryptosslRSA.hh
-         XrdCrypto/openssl3/XrdCryptosslX509.cc      XrdCrypto/XrdCryptosslX509.hh
-         XrdCrypto/XrdCryptosslX509Crl.cc            XrdCrypto/XrdCryptosslX509Crl.hh
-         XrdCrypto/XrdCryptosslX509Req.cc            XrdCrypto/XrdCryptosslX509Req.hh
-         XrdCrypto/XrdCryptosslTrace.hh
-         XrdCrypto/openssl3/XrdCryptosslFactory.cc   XrdCrypto/XrdCryptosslFactory.hh )
-  else()
-    set( XrdCryptosslSources
-         XrdCrypto/XrdCryptosslAux.cc            XrdCrypto/XrdCryptosslAux.hh
-         XrdCrypto/XrdCryptosslgsiAux.cc
-         XrdCrypto/XrdCryptosslCipher.cc         XrdCrypto/XrdCryptosslCipher.hh
-         XrdCrypto/XrdCryptosslMsgDigest.cc      XrdCrypto/XrdCryptosslMsgDigest.hh
-         XrdCrypto/XrdCryptosslRSA.cc            XrdCrypto/XrdCryptosslRSA.hh
-         XrdCrypto/XrdCryptosslX509.cc           XrdCrypto/XrdCryptosslX509.hh
-         XrdCrypto/XrdCryptosslX509Crl.cc        XrdCrypto/XrdCryptosslX509Crl.hh
-         XrdCrypto/XrdCryptosslX509Req.cc        XrdCrypto/XrdCryptosslX509Req.hh
-         XrdCrypto/XrdCryptosslTrace.hh
-         XrdCrypto/XrdCryptosslFactory.cc        XrdCrypto/XrdCryptosslFactory.hh )
-  endif()
+  set( XrdCryptosslSources
+       XrdCrypto/XrdCryptosslAux.cc       XrdCrypto/XrdCryptosslAux.hh
+       XrdCrypto/XrdCryptosslgsiAux.cc
+       XrdCrypto/XrdCryptosslCipher.cc    XrdCrypto/XrdCryptosslCipher.hh
+       XrdCrypto/XrdCryptosslMsgDigest.cc XrdCrypto/XrdCryptosslMsgDigest.hh
+       XrdCrypto/XrdCryptosslRSA.cc       XrdCrypto/XrdCryptosslRSA.hh
+       XrdCrypto/XrdCryptosslX509.cc      XrdCrypto/XrdCryptosslX509.hh
+       XrdCrypto/XrdCryptosslX509Crl.cc   XrdCrypto/XrdCryptosslX509Crl.hh
+       XrdCrypto/XrdCryptosslX509Req.cc   XrdCrypto/XrdCryptosslX509Req.hh
+       XrdCrypto/XrdCryptosslTrace.hh
+       XrdCrypto/XrdCryptosslFactory.cc   XrdCrypto/XrdCryptosslFactory.hh )
 
   add_library(
     ${LIB_XRD_CRYPTOSSL}
