+++ src/XrdCrypto/XrdCryptoLite_bf32.cc
@@ -40,7 +40,7 @@
 #include <netinet/in.h>
 #include <cinttypes>
 
-#include <openssl/blowfish.h>
+#include <openssl/evp.h>
 
 #include "XrdOuc/XrdOucCRC.hh"
 #include "XrdSys/XrdSysHeaders.hh"
@@ -82,23 +82,23 @@ int XrdCryptoLite_bf32::Decrypt(const char *key,
                                 char       *dst,
                                 int         dstLen)
 {
-   BF_KEY decKey;
    unsigned char ivec[8] = {0,0,0,0,0,0,0,0};
    unsigned int crc32;
-   int ivnum = 0, dLen = srcLen-sizeof(crc32);
+   int wLen;
+   int dLen = srcLen - sizeof(crc32);
 
 // Make sure we have data
 //
    if (dstLen <= (int)sizeof(crc32) || dstLen < srcLen) return -EINVAL;
 
-// Set the key
-//
-   BF_set_key(&decKey, keyLen, (const unsigned char *)key);
-
 // Decrypt
 //
-   BF_cfb64_encrypt((const unsigned char *)src, (unsigned char *)dst, srcLen,
-                    &decKey, ivec, &ivnum, BF_DECRYPT);
+   EVP_CIPHER_CTX *ctx = EVP_CIPHER_CTX_new();
+   EVP_DecryptInit_ex(ctx, EVP_bf_cfb64(), 0, (unsigned char *)key, ivec);
+   EVP_DecryptUpdate(ctx, (unsigned char *)dst, &wLen,
+                          (unsigned char *)src, srcLen);
+   EVP_DecryptFinal_ex(ctx, (unsigned char *)dst, &wLen);
+   EVP_CIPHER_CTX_free(ctx);
 
 // Perform the CRC check to verify we have valid data here
 //
@@ -123,10 +123,10 @@ int XrdCryptoLite_bf32::Encrypt(const char *key,
                                 char       *dst,
                                 int         dstLen)
 {
-   BF_KEY encKey;
    unsigned char buff[4096], *bP, *mP = 0, ivec[8] = {0,0,0,0,0,0,0,0};
    unsigned int crc32;
-   int ivnum = 0, dLen = srcLen+sizeof(crc32);
+   int wLen;
+   int dLen = srcLen + sizeof(crc32);
 
 // Make sure that the destination if at least 4 bytes larger and we have data
 //
@@ -146,14 +146,13 @@ int XrdCryptoLite_bf32::Encrypt(const char *key,
    crc32 = htonl(crc32);
    memcpy((bP+srcLen), &crc32, sizeof(crc32));
 
-// Set the key
-//
-   BF_set_key(&encKey, keyLen, (const unsigned char *)key);
-
 // Encrypt
 //
-   BF_cfb64_encrypt(bP, (unsigned char *)dst, dLen,
-                    &encKey, ivec, &ivnum, BF_ENCRYPT);
+   EVP_CIPHER_CTX *ctx = EVP_CIPHER_CTX_new();
+   EVP_EncryptInit_ex(ctx, EVP_bf_cfb64(), 0, (unsigned char *)key, ivec);
+   EVP_EncryptUpdate(ctx, (unsigned char *)dst, &wLen, bP, dLen);
+   EVP_EncryptFinal_ex(ctx, (unsigned char *)dst, &wLen);
+   EVP_CIPHER_CTX_free(ctx);
 
 // Free temp buffer and return success
 //
