+++ src/XrdSec.cmake
@@ -47,15 +47,9 @@ set_target_properties(
 #-------------------------------------------------------------------------------
 # The XrdSecpwd module
 #-------------------------------------------------------------------------------
-if( ENABLE_OPENSSL3 )
-  set( XrdSecProtectSources
-    XrdSec/openssl3/XrdSecProtect.cc    XrdSec/XrdSecProtect.hh
-    XrdSec/XrdSecProtector.cc           XrdSec/XrdSecProtector.hh )
-else()
-  set( XrdSecProtectSources
-    XrdSec/XrdSecProtect.cc             XrdSec/XrdSecProtect.hh
-    XrdSec/XrdSecProtector.cc           XrdSec/XrdSecProtector.hh )
-endif()
+set( XrdSecProtectSources
+     XrdSec/XrdSecProtect.cc             XrdSec/XrdSecProtect.hh
+     XrdSec/XrdSecProtector.cc           XrdSec/XrdSecProtector.hh )
 
 add_library(
   ${LIB_XRD_SEC_PROT}
