+++ src/XrdUtils.cmake
@@ -58,70 +58,39 @@ set ( XrdSysSources
   XrdSys/XrdSysXSLock.cc        XrdSys/XrdSysXSLock.hh
 )
 
-  #-----------------------------------------------------------------------------
-  # XrdTls
-  #-----------------------------------------------------------------------------
-if ( ENABLE_OPENSSL3 )
-  set ( XrdTlsSources
-    XrdTls/XrdTls.cc                  XrdTls/XrdTls.hh
-    XrdTls/openssl3/XrdTlsContext.cc  XrdTls/XrdTlsContext.hh
-    XrdTls/XrdTlsHostcheck.icc        XrdTls/XrdTlsHostcheck.hh
-    XrdTls/XrdTlsNotary.cc            XrdTls/XrdTlsNotary.hh
-    XrdTls/XrdTlsNotaryUtils.icc      XrdTls/XrdTlsNotaryUtils.hh
-    XrdTls/XrdTlsPeerCerts.cc         XrdTls/XrdTlsPeerCerts.hh
-    XrdTls/XrdTlsSocket.cc            XrdTls/XrdTlsSocket.hh
-    XrdTls/XrdTlsTempCA.cc            XrdTls/XrdTlsTempCA.hh
-  )
-else()
-  set ( XrdTlsSources
-    XrdTls/XrdTls.cc                  XrdTls/XrdTls.hh
-    XrdTls/XrdTlsContext.cc           XrdTls/XrdTlsContext.hh
-    XrdTls/XrdTlsHostcheck.icc        XrdTls/XrdTlsHostcheck.hh
-    XrdTls/XrdTlsNotary.cc            XrdTls/XrdTlsNotary.hh
-    XrdTls/XrdTlsNotaryUtils.icc      XrdTls/XrdTlsNotaryUtils.hh
-    XrdTls/XrdTlsPeerCerts.cc         XrdTls/XrdTlsPeerCerts.hh
-    XrdTls/XrdTlsSocket.cc            XrdTls/XrdTlsSocket.hh
-    XrdTls/XrdTlsTempCA.cc            XrdTls/XrdTlsTempCA.hh
-  )
-endif()
+#-----------------------------------------------------------------------------
+# XrdTls
+#-----------------------------------------------------------------------------
+set ( XrdTlsSources
+      XrdTls/XrdTls.cc                  XrdTls/XrdTls.hh
+      XrdTls/XrdTlsContext.cc           XrdTls/XrdTlsContext.hh
+      XrdTls/XrdTlsHostcheck.icc        XrdTls/XrdTlsHostcheck.hh
+      XrdTls/XrdTlsNotary.cc            XrdTls/XrdTlsNotary.hh
+      XrdTls/XrdTlsNotaryUtils.icc      XrdTls/XrdTlsNotaryUtils.hh
+      XrdTls/XrdTlsPeerCerts.cc         XrdTls/XrdTlsPeerCerts.hh
+      XrdTls/XrdTlsSocket.cc            XrdTls/XrdTlsSocket.hh
+      XrdTls/XrdTlsTempCA.cc            XrdTls/XrdTlsTempCA.hh
+)
 
-  #-----------------------------------------------------------------------------
-  # XrdCrypto: linking against a few functions that are useful for XrdTls; avoids
-  # linking against all of libXrdCryptossl in XrdUtils
-  #-----------------------------------------------------------------------------
-if ( ENABLE_OPENSSL3 )
-  set ( XrdCryptoSources
-    XrdCrypto/openssl3/XrdCryptosslAux.cc     XrdCrypto/XrdCryptosslAux.hh
-    XrdCrypto/openssl3/XrdCryptosslX509.cc    XrdCrypto/XrdCryptosslX509.hh
-    XrdCrypto/XrdCryptoX509.cc                XrdCrypto/XrdCryptoX509.hh
-    XrdCrypto/XrdCryptoX509Chain.cc           XrdCrypto/XrdCryptoX509Chain.hh
-    XrdCrypto/openssl3/XrdCryptosslRSA.cc     XrdCrypto/XrdCryptosslRSA.hh
-    XrdCrypto/XrdCryptoRSA.cc                 XrdCrypto/XrdCryptoRSA.hh
-    XrdCrypto/openssl3/XrdCryptosslgsiAux.cc  XrdCrypto/XrdCryptosslgsiAux.hh
-    XrdCrypto/XrdCryptosslX509Req.cc          XrdCrypto/XrdCryptosslX509Req.hh
-    XrdCrypto/XrdCryptoX509Req.cc             XrdCrypto/XrdCryptoX509Req.hh
-    XrdCrypto/XrdCryptoAux.cc                 XrdCrypto/XrdCryptoAux.hh
-    XrdCrypto/XrdCryptosslX509Crl.cc          XrdCrypto/XrdCryptosslX509Crl.hh
-    XrdCrypto/XrdCryptoX509Crl.cc             XrdCrypto/XrdCryptoX509Crl.hh
-                                              XrdCrypto/XrdCryptoTrace.hh
-  )
-else()
-  set ( XrdCryptoSources
-    XrdCrypto/XrdCryptosslAux.cc              XrdCrypto/XrdCryptosslAux.hh
-    XrdCrypto/XrdCryptosslX509.cc             XrdCrypto/XrdCryptosslX509.hh
-    XrdCrypto/XrdCryptoX509.cc                XrdCrypto/XrdCryptoX509.hh
-    XrdCrypto/XrdCryptoX509Chain.cc           XrdCrypto/XrdCryptoX509Chain.hh
-    XrdCrypto/XrdCryptosslRSA.cc              XrdCrypto/XrdCryptosslRSA.hh
-    XrdCrypto/XrdCryptoRSA.cc                 XrdCrypto/XrdCryptoRSA.hh
-    XrdCrypto/XrdCryptosslgsiAux.cc           XrdCrypto/XrdCryptosslgsiAux.hh
-    XrdCrypto/XrdCryptosslX509Req.cc          XrdCrypto/XrdCryptosslX509Req.hh
-    XrdCrypto/XrdCryptoX509Req.cc             XrdCrypto/XrdCryptoX509Req.hh
-    XrdCrypto/XrdCryptoAux.cc                 XrdCrypto/XrdCryptoAux.hh
-    XrdCrypto/XrdCryptosslX509Crl.cc          XrdCrypto/XrdCryptosslX509Crl.hh
-    XrdCrypto/XrdCryptoX509Crl.cc             XrdCrypto/XrdCryptoX509Crl.hh
-                                              XrdCrypto/XrdCryptoTrace.hh
-  )
-endif()
+#-----------------------------------------------------------------------------
+# XrdCrypto: linking against a few functions that are useful for XrdTls; avoids
+# linking against all of libXrdCryptossl in XrdUtils
+#-----------------------------------------------------------------------------
+set ( XrdCryptoSources
+      XrdCrypto/XrdCryptosslAux.cc              XrdCrypto/XrdCryptosslAux.hh
+      XrdCrypto/XrdCryptosslX509.cc             XrdCrypto/XrdCryptosslX509.hh
+      XrdCrypto/XrdCryptoX509.cc                XrdCrypto/XrdCryptoX509.hh
+      XrdCrypto/XrdCryptoX509Chain.cc           XrdCrypto/XrdCryptoX509Chain.hh
+      XrdCrypto/XrdCryptosslRSA.cc              XrdCrypto/XrdCryptosslRSA.hh
+      XrdCrypto/XrdCryptoRSA.cc                 XrdCrypto/XrdCryptoRSA.hh
+      XrdCrypto/XrdCryptosslgsiAux.cc           XrdCrypto/XrdCryptosslgsiAux.hh
+      XrdCrypto/XrdCryptosslX509Req.cc          XrdCrypto/XrdCryptosslX509Req.hh
+      XrdCrypto/XrdCryptoX509Req.cc             XrdCrypto/XrdCryptoX509Req.hh
+      XrdCrypto/XrdCryptoAux.cc                 XrdCrypto/XrdCryptoAux.hh
+      XrdCrypto/XrdCryptosslX509Crl.cc          XrdCrypto/XrdCryptosslX509Crl.hh
+      XrdCrypto/XrdCryptoX509Crl.cc             XrdCrypto/XrdCryptoX509Crl.hh
+                                                XrdCrypto/XrdCryptoTrace.hh
+)
 
   #-----------------------------------------------------------------------------
   # XrdOuc
